--- snapshots/Cargo.toml
@@ -1,32 +1,0 @@
-# Copyright 2018-2021 the Deno authors. All rights reserved. MIT license.
-
-# IMPORTANT(bartlomieju): this crate is internal and shouldn't be published
-# to crates.io
-
-[package]
-name = "deno_snapshots"
-version = "0.1.0"
-authors = ["the Deno authors"]
-edition = "2018"
-license = "MIT"
-repository = "https://github.com/denoland/deno"
-description = "Provides snapshots of TSC & Deno (runtime+web+core)"
-
-[lib]
-name = "deno_snapshots"
-path = "lib.rs"
-
-[dependencies]
-deno_core = { version = "0.141.0", path = "../core" } # For mock TSC #[op]s
-deno_runtime = { version = "0.67.0", path = "../runtime" }
-lzzzz = "1.0"
-once_cell = "1.10.0"
-zstd = "0.11.1"
-
-[build-dependencies]
-deno_core = { version = "0.141.0", path = "../core" } # For mock TSC #[op]s
-deno_runtime = { version = "0.67.0", path = "../runtime" }
-lzzzz = "1.0"
-regex = "1.5.6"
-serde = { version = "1.0.125", features = ["derive"] }
-zstd = "0.11.1"
